BZ-1292768 Message loss or duplicate over the bridge during outages
(cherry picked from commit 98f3a12373eb66a6b6f92f47392a7ddf37d84628)
(cherry picked from commit 35ece3ef07206c8f1e2e0fa45c5a14084eb53177)

BZ-1292768 Message loss or duplicate fix on shrinking duplicate cache

(cherry picked from commit 790f1427d70e162fb762a0f7b962380d9150eb29)
(cherry picked from commit 8af2cf1f64336e686f2761e592ed1ceaa9dc27d9)

BZ-1292768 Message loss or duplicate critical exceptions from memory

(cherry picked from commit b53a95482b397eab06a1ba87cb7277f5d0c6d883)
(cherry picked from commit 5b020cece8f3ba4dde8ea7f88a9c628da9dd1c8b)

+++ hornetq-server/src/main/java/org/hornetq/core/postoffice/impl/PostOfficeImpl.java
@@ -488,7 +488,7 @@ public class PostOfficeImpl implements PostOffice, NotificationListener, Binding
       managementService.sendNotification(new Notification(uid, CoreNotificationType.BINDING_ADDED, props));
    }
 
-   public synchronized Binding removeBinding(final SimpleString uniqueName, Transaction tx) throws Exception
+   public synchronized Binding removeBinding(final SimpleString uniqueName, Transaction tx, boolean deleteData) throws Exception
    {
 
       addressSettingsRepository.clearCache();
@@ -500,7 +500,7 @@ public class PostOfficeImpl implements PostOffice, NotificationListener, Binding
          throw new HornetQNonExistentQueueException();
       }
 
-      if (addressManager.getBindingsForRoutingAddress(binding.getAddress()) == null)
+      if (deleteData && addressManager.getBindingsForRoutingAddress(binding.getAddress()) == null)
       {
          pagingManager.deletePageStore(binding.getAddress());
 
@@ -1245,32 +1245,21 @@ public class PostOfficeImpl implements PostOffice, NotificationListener, Binding
 
          DuplicateIDCache cacheBridge = getDuplicateIDCache(BRIDGE_CACHE_STR.concat(message.getAddress()));
 
-         if (cacheBridge.contains(bridgeDupBytes))
+         if (context.getTransaction() == null)
          {
-            HornetQServerLogger.LOGGER.duplicateMessageDetectedThruBridge(message);
+            context.setTransaction(new TransactionImpl(storageManager));
+            startedTX.set(true);
+         }
 
-            if (context.getTransaction() != null)
-            {
-               context.getTransaction().markAsRollbackOnly(new HornetQDuplicateIdException());
-            }
-
+         if (!cacheBridge.atomicVerify(bridgeDupBytes, context.getTransaction()))
+         {
+            context.getTransaction().rollback();
+            startedTX.set(false);
             message.decrementRefCount();
-
             return false;
          }
-         else
-         {
-            if (context.getTransaction() == null)
-            {
-               context.setTransaction(new TransactionImpl(storageManager));
-               startedTX.set(true);
-            }
-         }
-
-         cacheBridge.addToCache(bridgeDupBytes, context.getTransaction());
 
          message.removeProperty(MessageImpl.HDR_BRIDGE_DUPLICATE_ID);
-
       }
       else
       {
@@ -1315,7 +1304,7 @@ public class PostOfficeImpl implements PostOffice, NotificationListener, Binding
                startedTX.set(true);
             }
 
-            cache.addToCache(duplicateIDBytes, context.getTransaction());
+            cache.addToCache(duplicateIDBytes, context.getTransaction(), false);
          }
       }
